Update README.md
The Lord of the Rings* series began as a novel by **J.R.R. Tolkien**, first published in 1954–1955. Initially intended as a sequel to his children's book, *The Hobbit*, it evolved into a much larger work, deeply influenced by Tolkien’s love for languages, myth, and ancient history. The success of the books solidified Tolkien as the father of modern fantasy literature.